requirements.txt
Django==3.2.18
djangorestframework==3.14.0
django-cors-headers==3.13.0
django-guardian==2.4.0
psycopg2-binary==2.9.6
python-decouple==3.8
djangorestframework-simplejwt==5.2.2
ipython==8.12.1
django-debug-toolbar==3.6.0
drf-yasg==1.20.0
phonenumbers==8.13.20
djangorestframework>=3.0.0
drf-yasg>=1.0.0
